# app-backend/docs/swagger.yaml
basePath: /
definitions:
...
  p2p.depositEscrowRequest:
    properties:
      commit:
        type: integer
      transaction:
        type: string
      transactionSignature:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Core wallet API for the Trovo platform — user accounts, crypto assets,
    deposits/withdrawals, swaps.
  title: Trovo Wallet API
  version: "1.0"
paths:
...
  /v1/p2p/orders/{orderID}/escrow-deposit:
    post:
      consumes:
      - application/json
      description: 'Used by the asset depositor (the SELL-side party) to fund escrow
        for an order. Call it twice: first with an empty body to receive an unsigned
        transaction to sign locally, then again with the signed transaction and commit=1
        to actually submit it. The source wallet must be a standard wallet with no
        approver-based shared access.'
      parameters:
      - description: Order ID
        in: path
        name: orderID
        required: true
        type: string
      - description: transaction (base64, from a prior unsigned call), transactionSignature,
          commit (1 to submit)
        in: body
        name: body
        schema:
          $ref: '#/definitions/p2p.depositEscrowRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Unsigned transaction to sign, or submission result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not the asset depositor, wrong wallet type, or shared-access
            wallet with approvers
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Deposit the sold asset into P2P escrow
      tags:
      - P2P
  /v1/p2p/orders/{orderID}/escrow-deposit/regenerate-shortlink:
    post:
      description: Retries shortlink generation if the automatic attempt made when
        the order was accepted failed and left the order without a deposit shortlink.
      parameters:
      - description: Order ID
        in: path
        name: orderID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Regenerate the shareable escrow-deposit link for an order
      tags:
      - P2P
...
securityDefinitions:
  BearerAuth:
    description: JWT bearer token used by the internal "trovo-manager" (tm-api admin)
      routes. Send as "Bearer <token>". See internal/middleware/authentication_middleware.go's
      JwtTokenAuthMiddleware.
    in: header
    name: Authorization
    type: apiKey
  ServiceLinkApiKey:
    description: Static per-partner API key used by white-label "service link" integrations
      (and by tm-api for a few admin/server-to-server calls). Issued when a service
      link is provisioned; see internal/components/servicelinks. See internal/middleware/api_key_middleware.go.
    in: header
    name: X-TW-SERVICE-LINK-API-KEY
    type: apiKey
  SignatureAuth:
    description: 'Request-signing scheme used by app-web and app-mobile for end-user
      requests. The caller sends four headers together: X-TW-PUBLIC-KEY (the wallet
      address making the call), X-TW-SIGNER (the address whose private key signed
      the request), X-TW-TIMESTAMP (a fresh timestamp), and X-TW-SIGNATURE (a signature
      over signer+timestamp, and for POST/PUT/PATCH over the request body/URI). There
      is no bearer token: every request is independently signed. See internal/middleware/authentication_middleware.go.'
    in: header
    name: X-TW-SIGNATURE
    type: apiKey
swagger: "2.0"